requirements.txt
click==7.1.2
configparser==4.0.2
contextlib2==0.6.0.post1
Flask==1.1.4
flask_restful
Flask-SQLAlchemy==2.5.1
importlib-metadata==2.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==2.21.0
mysql-connector-python==8.0.27
pathlib2==2.3.6
protobuf==3.17.3
scandir==1.10.0
six==1.16.0
SQLAlchemy==1.4.26
Werkzeug==1.0.1
zipp==1.2.0
